fix: meal entity, dto 필드 수정

+++ back/healthone-meal-service/src/main/java/com/secui/healthone/domain/meal/dto/MealResponseDto.java
@@ -20,7 +20,7 @@ public class MealResponseDto {
     private MealType mealType;
     private Float portion;
     private Float gram;
-    private Float kcal;
+    private Integer kcal;
 
     @Builder
     public MealResponseDto(Meal entity) {
